Move IToolUseSession from implementations to agent/core/flows
CycleServices in agent/core imported IToolUseSession from
agent/implementations, inverting the documented dependency rule
("Dependencies point inward, never the reverse" from core/README.md).

Extract the interface into src/agent/core/flows/IToolUseSession.ts so
core code can reference it without depending on implementations. Update
ToolUseSessionLifecycle to import and implement the interface from core;
update ToolUseServices and the tooluse barrel to source it from core as
well.
